WIP: Debuggint test
