--- spec/integration/dashboard/graphs_spec.js
@@ -1,11 +1,0 @@
-describe("dashboard graphs", () => {
-  before(() => {
-    cy.visit('./');
-  });
-
-  context("default settings", () => {
-    it('The title of the graph should match the text of the button', () => {
-      // cy.get('')
-    });
-  })
-});
