+++ src/lib/formatters/DateFormatters.js
@@ -1,0 +1,5 @@
+const shortDate = (date) => {
+   return new Intl.DateTimeFormat('en-US', {month:'short'}).format(new Date(date))
+}
+
+export default { shortDate }
