+++ src/spec/unit/NavBar/NavBar.spec.js
@@ -1,0 +1,26 @@
+describe('NavBar behavior', () => {
+    before(() => {
+        cy.visit('/')
+    });
+    context('When interacting with the links', () => {
+        it('The current view is described by the active button', () => {
+
+        });
+        it('Only one link is active', () => {
+            
+        });
+        describe('and I click a link:', () => {
+            it('The app moves to a new location', () => {
+
+            });
+            it('THe active class is applied to the clicked link', () => {
+
+            });
+        });
+    });
+    context('When interacting with the type bar', ()=> {
+        it('will render a modal if enter is pressed', () => {
+
+        });
+    });
+});
